+++ src/test/java/services/sandbox/MorningDigestSchedulerTest.java
@@ -1,5 +1,6 @@
 package services.sandbox;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
 import services.sandbox.model.Position;
 import services.sandbox.model.SandboxUser;
@@ -35,8 +36,8 @@ public class MorningDigestSchedulerTest {
         SandboxTradingService tradingService = mock(SandboxTradingService.class);
         MorningDigestScheduler sched = scheduler(tradingService);
 
-        SandboxUser user = new SandboxUser("u1", "Alice", 100_000.0);
-        Position closedPos = new Position("u1", "SBER", "uid-sber", 0, 300.0);
+        SandboxUser user = new SandboxUser("u1", "Alice", BigDecimal.valueOf(100_000.0));
+        Position closedPos = new Position("u1", "SBER", "uid-sber", 0, BigDecimal.valueOf(300.0));
 
         String digest = sched.buildDigest(user, List.of(closedPos));
 
@@ -49,8 +50,8 @@ public class MorningDigestSchedulerTest {
         when(tradingService.price("GAZP")).thenReturn("GAZP = 180 ₽");
         MorningDigestScheduler sched = scheduler(tradingService);
 
-        SandboxUser user = new SandboxUser("u1", "Alice", 50_000.0);
-        Position openPos = new Position("u1", "GAZP", "uid-gazp", 10, 175.0);
+        SandboxUser user = new SandboxUser("u1", "Alice", BigDecimal.valueOf(50_000.0));
+        Position openPos = new Position("u1", "GAZP", "uid-gazp", 10, BigDecimal.valueOf(175.0));
 
         String digest = sched.buildDigest(user, List.of(openPos));
