+++ src/test/java/services/sandbox/SandboxTradingLogicTest.java
@@ -47,17 +47,17 @@ public class SandboxTradingLogicTest {
 		int week = now.get(WeekFields.ISO.weekOfWeekBasedYear());
 		if (u.getDailyBaselineDate() == null || !now.equals(u.getDailyBaselineDate())) {
 			u.setDailyBaselineDate(now);
-			u.setDailyBaselineEquity(equityAtCallTime.doubleValue());
+			u.setDailyBaselineEquity(equityAtCallTime);
 		}
 		if (u.getWeeklyBaselineDate() == null ||
 				u.getWeeklyBaselineDate().get(WeekFields.ISO.weekOfWeekBasedYear()) != week) {
 			u.setWeeklyBaselineDate(now);
-			u.setWeeklyBaselineEquity(equityAtCallTime.doubleValue());
+			u.setWeeklyBaselineEquity(equityAtCallTime);
 		}
 		if (u.getMonthlyBaselineDate() == null ||
 				u.getMonthlyBaselineDate().getMonthValue() != now.getMonthValue()) {
 			u.setMonthlyBaselineDate(now);
-			u.setMonthlyBaselineEquity(equityAtCallTime.doubleValue());
+			u.setMonthlyBaselineEquity(equityAtCallTime);
 		}
 	}
 
@@ -73,18 +73,18 @@ public class SandboxTradingLogicTest {
 	 */
 	@Test
 	void testRecordBaseline_sameDay_doesNotOverwriteExistingBaseline() {
-		SandboxUser user = new SandboxUser("u1", "Alice", 100000.00);
+		SandboxUser user = new SandboxUser("u1", "Alice", BigDecimal.valueOf(100000.00));
 		LocalDate today = LocalDate.of(2026, 3, 19);
 
 		// First call (e.g. from register) — equity = 100 000
 		recordBaseline(user, new BigDecimal("100000.00"), today);
-		assertEquals(0, new BigDecimal("100000.00").compareTo(BigDecimal.valueOf(user.getDailyBaselineEquity())),
+		assertEquals(0, new BigDecimal("100000.00").compareTo(user.getDailyBaselineEquity()),
 				"Daily baseline should be set to 100 000 on first call");
 
 		// After a profitable trade equity grows to 110 000.
 		// A second call on the same day should NOT overwrite the baseline.
 		recordBaseline(user, new BigDecimal("110000.00"), today);
-		assertEquals(0, new BigDecimal("100000.00").compareTo(BigDecimal.valueOf(user.getDailyBaselineEquity())),
+		assertEquals(0, new BigDecimal("100000.00").compareTo(user.getDailyBaselineEquity()),
 				"Daily baseline must NOT be updated when the date hasn't changed");
 	}
 
@@ -94,13 +94,13 @@ public class SandboxTradingLogicTest {
 	 */
 	@Test
 	void testRecordBaseline_newDay_updatesBaselineToPreTradeEquity() {
-		SandboxUser user = new SandboxUser("u1", "Alice", 100000.00);
+		SandboxUser user = new SandboxUser("u1", "Alice", BigDecimal.valueOf(100000.00));
 		LocalDate yesterday = LocalDate.of(2026, 3, 18);
 		LocalDate today = LocalDate.of(2026, 3, 19);
 
 		// Baseline was set yesterday
 		recordBaseline(user, new BigDecimal("105000.00"), yesterday);
-		assertEquals(0, new BigDecimal("105000.00").compareTo(BigDecimal.valueOf(user.getDailyBaselineEquity())));
+		assertEquals(0, new BigDecimal("105000.00").compareTo(user.getDailyBaselineEquity()));
 
 		// Today's first trade: recordBaseline() is called BEFORE the trade,
 		// so equityAtCallTime = equity before the trade = 107 000.
@@ -108,7 +108,7 @@ public class SandboxTradingLogicTest {
 		recordBaseline(user, equityBeforeTrade, today);
 
 		assertEquals(today, user.getDailyBaselineDate());
-		assertEquals(0, new BigDecimal("107000.00").compareTo(BigDecimal.valueOf(user.getDailyBaselineEquity())),
+		assertEquals(0, new BigDecimal("107000.00").compareTo(user.getDailyBaselineEquity()),
 				"Baseline must capture equity BEFORE the trade on the new day");
 	}
 
@@ -136,17 +136,17 @@ public class SandboxTradingLogicTest {
 	 */
 	@Test
 	void testRecordBaseline_newWeek_updatesWeeklyBaseline() {
-		SandboxUser user = new SandboxUser("u1", "Alice", 100000.00);
+		SandboxUser user = new SandboxUser("u1", "Alice", BigDecimal.valueOf(100000.00));
 
 		// Week 11 of 2026
 		LocalDate lastWeek = LocalDate.of(2026, 3, 9);   // Monday, week 11
 		LocalDate thisWeek = LocalDate.of(2026, 3, 16);  // Monday, week 12
 
 		recordBaseline(user, new BigDecimal("100000.00"), lastWeek);
-		assertEquals(0, new BigDecimal("100000.00").compareTo(BigDecimal.valueOf(user.getWeeklyBaselineEquity())));
+		assertEquals(0, new BigDecimal("100000.00").compareTo(user.getWeeklyBaselineEquity()));
 
 		recordBaseline(user, new BigDecimal("115000.00"), thisWeek);
-		assertEquals(0, new BigDecimal("115000.00").compareTo(BigDecimal.valueOf(user.getWeeklyBaselineEquity())),
+		assertEquals(0, new BigDecimal("115000.00").compareTo(user.getWeeklyBaselineEquity()),
 				"Weekly baseline must update on a new ISO week");
 	}
